feat: add snapshot creator
